if, else, total marks
Write a JavaScript program to check the total marks of a student in various examinations. The student will get A+ grade if the total marks are in the range 89..100 inclusive, if the examination is "Final-exam." the student will get A+ grade where total marks must be greater than or equal to 90. Return true if the student get A+ grade or false otherwise.

// final이라는 단서와 is_exam의 매개변수가 어떻게 연결이 되는 걸까